[CI] Trim the base-c 4-gpu-h100 stage from 5 shards to 4 (#35407)
