+++ test/registered/hicache/test_hicache_storage_3fs_backend.py
@@ -13,7 +13,7 @@ from test_hicache_storage_file_backend import HiCacheStorageBaseMixin
 from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
 from sglang.test.test_utils import CustomTestCase
 
-register_cuda_ci(est_time=300, stage="base-c", runner_config="4-gpu-h100")
+register_cuda_ci(est_time=300, stage="base-b", runner_config="2-gpu-large")
 register_amd_ci(est_time=300, suite="base-b-test-2-gpu-large")
